Fix androidTest imports after modularization
Fixed imports in :app androidTest files to use correct module packages:

**TestDataBuilder.kt:**
- Updated to return Room entities (RoomExercise, RoomExerciseSet, RoomWorkoutPlan) instead of domain models
- DAOs expect Room entities, not domain models

**ExercisePaginationStabilityTest.kt:**
- Fixed RefittedRoomProvider import: `data.room` → `room`
- Added missing Exercise composable import from `ui.compose.exercise`

**TestRefittedRoomModule.kt:**
- Fixed RefittedRoom import: `data.room` → `room`
- Fixed RefittedRoomProvider import: `data.room` → `room`

These imports were outdated from before the modularization work. The androidTest code now compiles successfully and buildHealth runs without errors.
